[td-255] fix compiler error.

+++ src/query/src/qAggMain.c
@@ -4701,7 +4701,7 @@ static void irate_function_f(SQLFunctionCtx *pCtx, int32_t index) {
   SRateInfo *pRateInfo  = (SRateInfo *)GET_ROWCELL_INTERBUF(pResInfo);
   TSKEY     *primaryKey = GET_TS_LIST(pCtx);
   
-  int64_t v = 0;
+  double v = 0;
   GET_TYPED_DATA(v, double, pCtx->inputType, pData);
 
   pRateInfo->firstKey   = pRateInfo->lastKey;
